disable previews on fed server for tests tagged @disablePreviews

+++ tests/acceptance/features/bootstrap/WebUIGeneralContext.php
@@ -56,7 +56,7 @@ class WebUIGeneralContext extends RawMinkContext implements Context {
 	private $loginPage;
 
 	private $oldCSRFSetting = null;
-	private $oldPreviewSetting = null;
+	private $oldPreviewSetting = [];
 	private $createdFiles = [];
 
 	/**
@@ -641,14 +641,18 @@ class WebUIGeneralContext extends RawMinkContext implements Context {
 	 * @throws \Exception
 	 */
 	public function disablePreviewBeforeScenario() {
-		if ($this->oldPreviewSetting === null) {
-			$oldPreviewSetting = $this->featureContext->getSystemConfigValue(
-				'enable_previews'
-			);
-			$this->oldPreviewSetting = \trim($oldPreviewSetting);
-		}
-		$this->featureContext->setSystemConfig(
-			'enable_previews', 'false', 'boolean'
+		$this->featureContext->runFunctionOnEveryServer(
+			function ($server) {
+				if (!isset($this->oldPreviewSetting[$server])) {
+					$oldPreviewSetting = $this->featureContext->getSystemConfigValue(
+						'enable_previews'
+					);
+					$this->oldPreviewSetting[$server] = \trim($oldPreviewSetting);
+				}
+				$this->featureContext->setSystemConfig(
+					'enable_previews', 'false', 'boolean'
+				);
+			}
 		);
 	}
 
@@ -665,14 +669,18 @@ class WebUIGeneralContext extends RawMinkContext implements Context {
 	 * @throws \Exception
 	 */
 	public function enablePreviewBeforeScenario() {
-		if ($this->oldPreviewSetting === null) {
-			$oldPreviewSetting = $this->featureContext->getSystemConfigValue(
-				'enable_previews'
-			);
-			$this->oldPreviewSetting = \trim($oldPreviewSetting);
-		}
-		$this->featureContext->setSystemConfig(
-			'enable_previews', 'true', 'boolean'
+		$this->featureContext->runFunctionOnEveryServer(
+			function ($server) {
+				if (!isset($this->oldPreviewSetting[$server])) {
+					$oldPreviewSetting = $this->featureContext->getSystemConfigValue(
+						'enable_previews'
+					);
+					$this->oldPreviewSetting[$server] = \trim($oldPreviewSetting);
+				}
+				$this->featureContext->setSystemConfig(
+					'enable_previews', 'true', 'boolean'
+				);
+			}
 		);
 	}
 
@@ -702,13 +710,19 @@ class WebUIGeneralContext extends RawMinkContext implements Context {
 			$this->savedCapabilitiesChanges
 		);
 
-		if ($this->oldPreviewSetting === "") {
-			$this->featureContext->deleteSystemConfig('enable_previews');
-		} elseif ($this->oldPreviewSetting !== null) {
-			$this->featureContext->setSystemConfig(
-				'enable_previews', $this->oldPreviewSetting, 'boolean'
-			);
-		}
+		$this->featureContext->runFunctionOnEveryServer(
+			function ($server) {
+				if (isset($this->oldPreviewSetting[$server])
+					&& $this->oldPreviewSetting[$server] === ""
+				) {
+					$this->featureContext->deleteSystemConfig('enable_previews');
+				} elseif (isset($this->oldPreviewSetting[$server])) {
+					$this->featureContext->setSystemConfig(
+						'enable_previews', $this->oldPreviewSetting[$server], 'boolean'
+					);
+				}
+			}
+		);
 		
 		if ($this->oldCSRFSetting === "") {
 			$this->featureContext->deleteSystemConfig('csrf.disabled');
